Object Versioning: source transformations to proxy literal objects
Object Creation Point 1: creating objects through literals

for this versioning approach to work, objects need to wrapped by proxies.
that is, proxies need to be created for all (new) objects. for this,
source code should be transformed so that all literal objects are wrapped
into proxy-creation functions.

literals that create objects include literals for objects, for arrays,
and for functions.

numbers, strings, and boolean values are primitive.