Object Versioning: v8 benchmarks as integration tests
* adds two benchmarks from v8's benchmark suite: the deltaBlue constraint solver and
the Richards benchmark, which emulates an OS scheduler
* adds the test class IntegrationTestBenchmarks, which currently only includes a helper
method that requests the javascript file for a given URL, transforms the loaded JS source,
and then evaluates the transformed sources.

The idea is to use established JavaScript benchmarks both as integration tests and as
benchmarks. and also, to have continuous benchmark results while improving the
implementation of our approach, especially the global data structure, the object table.