Object Versioning: source transformations using UglifyJS's parser
further adds test cases to specify the different transformations of function
declarations  and function expressions, as it's not enough to simply wrap function
in the proxy function, but necessary to make the previously function declaration
(that is now an argument, that is, a function expressions) accessible via its name.

for more details see UglifyTransformer.transformSource() and its comments.

UglifyJS is able to also generate source maps, can also minify/beautify,
seems more complete and tested (e.g., doesn't strip function names in some
cases), doesn't throw exceptions during parsing, and is a magnitude faster
than the current Lively Parser for our transformations (and the current
benchmarks).