Object Versioning: simpler >>construct + .prototype handling
only emulate the new-operator's actions in the construct-trap to not require a
wrapper around the constructor (which was previously necessary as it's not possible
to supply a variable number of arguments when applying the new-operator):

1. create a new object (with Object.prototype as __proto__)
2. have that new object refer its actual prototype only through its __protoID
property
3. proxy the new object to have it use its __protoID property to delegate to its
__proto__-type(s!)
4. apply the original constructor with the new instance as this-argument and the
original args

when a function with a prototype property gets proxied, the prototype gets proxied
as well to have it be resolved via ID as well and to have all its non-primitive
properties also be referenced by IDs. having an ID for every prototype-property
also makes the construct-trap simpler.