Object Versioning: don't wrap global proxy, but transform evals
eval usually has access to the variable bindings of the function it
is called from. thus, code that gets evaluated can read local variables
and can define local variables. but when we wrap eval in a new function
as, for example, to first pass the code snippet through our source
transformation, code gets evaluated in the wrong scope and variables
defined in Global/window.

for this, reason we instead transform calls to the global eval and wrap
only the parameter into a call that transforms the code snippet
appropriately. therefore, as eval is no longer globally wrapped and
only source transformations therefore only active for eval's in transformed
code, we need to transform modules explicitly while loading them and also
explicitly patch all other relevant evals.