Object Versioning: transform instanceof to Object.instanceOf
there's no trap for the instanceOf operator, yet. instead it always
works directly on the proxy's original target (dummyTarget for our
fully virtual object-proxies), which might by now have a changed
prototype-chain.

further, a function's prototype property is proxied, whereas all
objects in the __proto__-chain of an object aren't proxied.. that is,
even resolving the __objectID to the current version of an object would
not be an instanceof its constructor..