Object Versioning: wrap global Objects/Constructors in proxyFor()
source transformation to wrap Array, Object, Function, Date, JSON,...
into lively.proxyFor()-calls to have them return proxies for new objects,
regardless of whether the objects are created using e.g. new Object(), or
just Object(), or Object.create().

as this change now also properly proxies Arrays in the benchmark, we reduced
the input sizes to the benchmarks to compensate for the immense slow-down...
just to not wait for ever..