Object Versioning: hoist func declarations after transformations
we transform
    function funcName(..) {..}
to
    var funcName = lively.proxyFor(function funcName(..) {..})

to have the function wrapped but still available by name in the current
scope. however, function declarations get hoisted and while variable
declarations get hoisted as well, the assignment doesn't. we, therefore,
explicitly move the var-assignment statements we create for function declarations to the beginning of scopes.