Object Versioning: simpler and sufficient redraw at undo/redo
lively's morphs render using HTML elements and the DOM. we chose not to version
DOM objects (Nodes, CSSStyleDeclarations) but instead only the morphs and their
internal graphical objects (shapes, renderContexts), because the DOM nodes are
also referred by the actual DOM tree and we would have to update on version
changes which nodes should be currently in the tree (by updating the parent
relations both on the previous and the current version of such nodes).

to keep things simple for this prototype, we instead chose to throw away and
recreate all DOM elements for the current lively world. this way, the versioning
of worlds is also more equivalent to serialization of worlds in that only the
objects of our graphics abstractions (=the morphs) get versioned and not the
output of the current rendering technology (=HTML elements generated by the
current renderContext that could also be, for example, SVG)

we don't just redraw the submorphs of the current lively world as in the
previous version, because morphs that where submorphs in the previous
version (before an undo/redo or setting a specific version) would still be
part of DOM tree..